// src/generated/models/Pulse_appsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pulse_appspulse_apptype = {
  895930000: 'Web',
  895930001: 'Mobile',
  895930002: 'Desktop'
} as const;
export type Pulse_appspulse_apptype = keyof typeof Pulse_appspulse_apptype;
export const Pulse_appspulse_opentype = {
  895930000: 'New Tab',
  895930001: 'Same Tab',
  895930002: 'Popup'
} as const;
export type Pulse_appspulse_opentype = keyof typeof Pulse_appspulse_opentype;
export const Pulse_appsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pulse_appsstatecode = keyof typeof Pulse_appsstatecode;
export const Pulse_appsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pulse_appsstatuscode = keyof typeof Pulse_appsstatuscode;

export interface Pulse_appsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pulse_appid: string;
  pulse_apptype?: Pulse_appspulse_apptype;
  pulse_appurl?: string;
  pulse_description?: string;
  pulse_iconurl?: string;
  pulse_isactive?: boolean;
  "pulse_Module@odata.bind"?: string;
  pulse_name?: string;
  pulse_opentype?: Pulse_appspulse_opentype;
  pulse_order?: number;
  statecode: Pulse_appsstatecode;
  statuscode?: Pulse_appsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pulse_apps extends Pulse_appsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pulse_apptypename?: string;
  pulse_isactivename?: string;
  pulse_modulename?: string;
  pulse_opentypename?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pulse_module?: object;
  _pulse_module_value?: string;
}